pyproject.toml
[project]
name = "financechallenge"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "marimo>=0.19.4",
    "matplotlib>=3.10.8",
    "neptune>=1.14.0",
    "numpy>=2.2.6",
    "optuna>=4.7.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "tqdm>=4.67.1",
]
